Rename InlineAssignEdit and remove save functionality
